COMPONENTS_EXAMPLE/requirements.txt
streamlit-aggrid
streamlit-plotly-events
streamlit-lottie
streamlit-pandas-profiling
streamlit-folium
streamlit-extras
